requirements.txt
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
yfinance>=0.2.31
python-dotenv>=1.0.0
fredapi>=0.5.1
python-telegram-bot>=20.7
scipy>=1.11.0
